requirements.txt
google-generativeai
dotenv
email_validator
Pillow
captcha
streamlit-js-eval
python-Levenshtein
azure-cognitiveservices-speech
pydub
audioop-lts
streamlit-webrtc
streamlit-drawable-canvas
ultralytics